% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR_function.R
\name{BMR}
\alias{BMR}
\title{Calculates Basal Metabolic Rate}
\usage{
BMR(pounds)
}
\arguments{
\item{pounds}{persons weight in pounds}
}
\value{
estimate of persons Basal Metabolic Rate
}
\description{
Calculates Basal Metabolic Rate
}
\examples{
# BMR Calculation
BMR(300)
}
