% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_Rawfiles.R
\name{clean_BMI}
\alias{clean_BMI}
\title{clean_BMI - rounds weight to the hundreths place and removes kg columns}
\usage{
clean_BMI(BMI_column)
}
\arguments{
\item{BMI_column}{BMI column}
}
\value{
cleaned BMI column
}
\description{
clean_BMI - rounds weight to the hundreths place and removes kg columns
}
