% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_Rawfiles.R
\name{clean_distance}
\alias{clean_distance}
\title{cleans raw data from inst/extdata for easier and cleaner package use}
\usage{
clean_distance(distance_column)
}
\arguments{
\item{distance_column}{cleans distance column by rounding}
}
\value{
cleaned distance column
}
\description{
clean_distance - rounds the distance to the hundreths place
}
