% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_Rawfiles.R
\name{clean_weight}
\alias{clean_weight}
\title{clean_weight - rounds weight to the hundreths place}
\usage{
clean_weight(data, kg_column, weight_column)
}
\arguments{
\item{data}{dataset}

\item{kg_column}{kg column to be removed}

\item{weight_column}{weight column}
}
\value{
cleaned weight column
}
\description{
clean_weight - rounds weight to the hundreths place
}
