% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{dailyActivity}
\alias{dailyActivity}
\title{Daily Activity Data}
\format{
A data frame with 10 variables:
\describe{
  \item{Id}{User ID}
  \item{ActivityDate}{Date of activity}
  \item{TotalSteps}{Total number of steps}
  \item{TotalDistance}{Total distance in miles}
  \item{DailyIntensity}{Aggregate intensity score}
  \item{activity_duration}{Duration of activity in minutes}
  \item{intense_activity_duration}{Time spent in intense activity (minutes)}
  \item{idle_duration}{Time spent idle (minutes)}
  \item{Calories}{Calories burned}
  \item{intensity_level}{Categorical intensity level}
}
}
\source{
Fitbit data
}
\usage{
dailyActivity
}
\description{
Contains information about users' daily physical activity.
}
\keyword{datasets}
