% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{dailySleep}
\alias{dailySleep}
\title{Daily Sleep Data}
\format{
A data frame with 5 variables:
\describe{
  \item{Id}{User ID}
  \item{SleepDate}{Date of recorded sleep}
  \item{TotalSleepRecords}{Number of sleep records in the night}
  \item{TotalMinutesAsleep}{Total minutes asleep}
  \item{TotalTimeInBed}{Total time spent in bed}
}
}
\source{
Fitbit data
}
\usage{
dailySleep
}
\description{
Contains information about users' sleep patterns.
}
\keyword{datasets}
