% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{dailyWeightLog}
\alias{dailyWeightLog}
\title{Daily Weight Log Data}
\format{
A data frame with 4 variables:
\describe{
  \item{Id}{User ID}
  \item{WeightDate}{Date of measurement}
  \item{WeightPounds}{Weight in pounds}
  \item{BMI}{Body Mass Index}
}
}
\source{
Fitbit data
}
\usage{
dailyWeightLog
}
\description{
Contains user weight and BMI records.
}
\keyword{datasets}
