% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{hourlyActivity}
\alias{hourlyActivity}
\title{Hourly Activity Data}
\format{
A data frame with 6 variables:
\describe{
  \item{Id}{User ID}
  \item{ActivityHour}{Timestamp for activity}
  \item{TotalSteps}{Steps taken during the hour}
  \item{TotalIntensity}{Intensity score for the hour}
  \item{Calories}{Calories burned}
  \item{intensity_level}{Categorical intensity level}
}
}
\source{
Fitbit data
}
\usage{
hourlyActivity
}
\description{
Contains hourly breakdown of user activity.
}
\keyword{datasets}
