% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_Rawfiles.R
\name{intensity_level}
\alias{intensity_level}
\title{Intensity levels}
\usage{
intensity_level(dataset, column)
}
\arguments{
\item{dataset}{dataset being modified}

\item{numeric_vector}{numeric column of activity levels}
}
\value{
updates dataset with categorical levels of activitity based off numeric column
}
\description{
Intensity levels
}
