% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_efficieny_function.R
\name{sleep_efficiency}
\alias{sleep_efficiency}
\title{Sleep Efficiency calculation}
\usage{
sleep_efficiency(minutesAsleep, timeInBed)
}
\arguments{
\item{minutesAsleep}{time spent asleep}

\item{timeInBed}{total time in bed}
}
\value{
sleep efficiency calculation
}
\description{
Sleep Efficiency calculation
}
\examples{
# Calculate Sleep Efficiency
sleep_efficiency(320,450)
}
