% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_visual.R
\name{sleep_plot}
\alias{sleep_plot}
\title{Sleep visual plot}
\usage{
sleep_plot(data, date, timeInBed, TimeAsleep, sleepRecorded)
}
\arguments{
\item{data}{dataset}

\item{date}{data column in dataset}

\item{timeInBed}{total time spent in bed}

\item{TimeAsleep}{time asleep}

\item{sleepRecorded}{number of times went to sleep}
}
\value{
plot of sleep data
}
\description{
Sleep visual plot
}
\examples{
# Plot sleep
sleep_plot(dailySleep, "SleepDate", "TotalTimeInBed", "TotalMinutesAsleep", "TotalSleepRecords")
}
