% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_visual.R
\name{steps_plot}
\alias{steps_plot}
\title{Step Visualization}
\usage{
steps_plot(data, date, steps)
}
\arguments{
\item{data}{dataset wanted for plot}

\item{date}{date column of dataset}

\item{steps}{step count of dataset}
}
\value{
ggplot of steps over time
}
\description{
Step Visualization
}
\examples{
# Plot steps
steps_plot(dailyActivity, "ActivityDate", "TotalSteps")
}
