% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_vs_calories_visual.R
\name{steps_vs_calories}
\alias{steps_vs_calories}
\title{Steps vs. Calories Plot}
\usage{
steps_vs_calories(data, steps, calories, intensity = NA, active_time = NA)
}
\arguments{
\item{data}{dataset}

\item{steps}{steps}

\item{calories}{calories burned}

\item{intensity}{optional: intensity of activity}

\item{active_time}{optional: total time of activity}
}
\value{
plot visualization
}
\description{
Steps vs. Calories Plot
}
\examples{
# Steps vs. Calories Plot
steps_vs_calories(dailyActivity, "TotalSteps", "Calories")
}
