% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDEE_function.R
\name{TDEE}
\alias{TDEE}
\title{Total Daily Energy Expenditure}
\usage{
TDEE(weight, intensity_level)
}
\arguments{
\item{weight}{weight in pounds}

\item{intensity}{character/category of intensity level}
}
\value{
TDEE calvulation
}
\description{
Total Daily Energy Expenditure
}
\examples{
# TDEE Plot
TDEE(230, "Low")
}
