% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_intake_function.R
\name{water_intake}
\alias{water_intake}
\title{Recommeded Water Intake Calculation}
\usage{
water_intake(weight, activityIntensity)
}
\arguments{
\item{weight}{weight}

\item{activityIntensity}{intensity "Low", "Medium", "High"}
}
\value{
recommended water intake in ounces calculations
}
\description{
Recommeded Water Intake Calculation
}
\examples{
# Recommended Water Intake Calculation
water_intake(230, "Low")
}
