% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{weekDaily}
\alias{weekDaily}
\title{Weekly Daily Aggregated Data}
\format{
A data frame with 10 variables:
\describe{
  \item{part_of_week}{Weekday or weekend}
  \item{day_of_week}{Day name (e.g., Monday)}
  \item{time_of_day}{Part of the day (e.g., morning, evening)}
  \item{total_steps}{Total steps in time block}
  \item{averaage_steps}{Average steps (typo preserved)}
  \item{total_intensities}{Total activity intensity}
  \item{average_intensities}{Average intensity score}
  \item{total_calories}{Total calories burned}
  \item{average_calories}{Average calories burned}
  \item{intensity_level}{Categorical intensity level}
}
}
\source{
Derived Fitbit data
}
\usage{
weekDaily
}
\description{
Aggregated metrics by part of the week and time of day.
}
\keyword{datasets}
