% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{weightLog}
\alias{weightLog}
\title{Weight Log Extended Data}
\format{
A data frame with 11 variables:
\describe{
  \item{Id}{User ID}
  \item{weight_pounds}{Weight in pounds}
  \item{BMI}{Body Mass Index}
  \item{body_type}{Categorical body type}
  \item{steps_count}{Associated step count}
  \item{distance_measured}{Distance in miles}
  \item{avg_intensities}{Average intensity score}
  \item{intense_activity_duration}{Duration of intense activity (min)}
  \item{active_duration}{Duration of activity (min)}
  \item{idle_duration}{Duration of idle time (min)}
  \item{intensity_level}{Categorical intensity level}
}
}
\source{
Derived Fitbit data
}
\usage{
weightLog
}
\description{
Enhanced user weight data with activity summaries.
}
\keyword{datasets}
